requirements.txt
python-gnupg==0.4.1
PyYAML==3.12